@@ -12,7 +12,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String debugLable = 'Unknown';
+  String? debugLable = 'Unknown';
   final JPush jpush = new JPush();
 
   @override
@@ -23,7 +23,7 @@ class _MyAppState extends State<MyApp> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
-    String platformVersion;
+    String? platformVersion;
 
     try {
       jpush.addEventHandler(
@@ -329,21 +329,20 @@ class _MyAppState extends State<MyApp> {
 
 /// 封装控件
 class CustomButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final String title;
+  final VoidCallback? onPressed;
+  final String? title;
 
   const CustomButton({@required this.onPressed, @required this.title});
 
   @override
   Widget build(BuildContext context) {
-    return new FlatButton(
+    return new TextButton(
       onPressed: onPressed,
       child: new Text("$title"),
-      color: Color(0xff585858),
-      highlightColor: Color(0xff888888),
-      splashColor: Color(0xff888888),
-      textColor: Colors.white,
-      //padding: EdgeInsets.fromLTRB(5, 5, 5, 5),
+      style: new ButtonStyle(foregroundColor: MaterialStateProperty.all(Colors.white),
+        overlayColor: MaterialStateProperty.all(Color(0xff888888)),
+        backgroundColor: MaterialStateProperty.all(Color(0xff585858)),
+        padding: MaterialStateProperty.all(EdgeInsets.fromLTRB(10, 5, 10, 5)), ),
     );
   }
 }
